+++ game/missiongenerator/aircraft/aircraftgenerator.py
@@ -35,6 +35,7 @@ from .aircraftpainter import AircraftPainter
 from .flightdata import FlightData
 from .flightgroupconfigurator import FlightGroupConfigurator
 from .flightgroupspawner import FlightGroupSpawner
+from ...data.weapons import WeaponType
 
 if TYPE_CHECKING:
     from game import Game
@@ -66,6 +67,10 @@ class AircraftGenerator:
         self.flights: List[FlightData] = []
         self.mission_data = mission_data
         self.helipads = helipads
+
+        self.ewrj_package_dict: Dict[int, List[FlyingGroup[Any]]] = {}
+        self.ewrj = settings.plugins.get("ewrj")
+        self.need_ecm = settings.plugin_option("ewrj.ecm_required")
 
     @cached_property
     def use_client(self) -> bool:
@@ -216,7 +221,24 @@ class AircraftGenerator:
             wpt.link_unit = hpad.id
             self.helipads[flight.arrival].units.append(hpad)
 
+        if self.ewrj:
+            self._track_ewrj_flight(flight, group)
+
         return group
+
+    def _track_ewrj_flight(self, flight: Flight, group: FlyingGroup[Any]) -> None:
+        if not self.ewrj_package_dict.get(id(flight.package)):
+            self.ewrj_package_dict[id(flight.package)] = []
+        if (
+            flight.package.primary_flight
+            and flight is flight.package.primary_flight
+            or flight.client_count
+            and (
+                not self.need_ecm
+                or flight.loadout.has_weapon_of_type(WeaponType.JAMMER)
+            )
+        ):
+            self.ewrj_package_dict[id(flight.package)].append(group)
 
     def _reserve_frequencies_and_tacan(self, ato: AirTaskingOrder) -> None:
         for package in ato.packages: